+++ core.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/osgi/RepairSchedulerService.java
@@ -16,13 +16,8 @@ package com.ericsson.bss.cassandra.ecchronos.core.osgi;
 
 import com.ericsson.bss.cassandra.ecchronos.core.JmxProxyFactory;
 import com.ericsson.bss.cassandra.ecchronos.core.TableStorageStates;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairConfiguration;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairJobView;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairLockType;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairScheduler;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.*;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.RepairStateFactory;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairSchedulerImpl;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.TableRepairJob;
 
 import com.ericsson.bss.cassandra.ecchronos.core.metrics.TableRepairMetrics;
 import com.ericsson.bss.cassandra.ecchronos.core.scheduling.ScheduleManager;
@@ -67,6 +62,10 @@ public class RepairSchedulerService implements RepairScheduler
     @Reference(service = TableStorageStates.class, cardinality = ReferenceCardinality.MANDATORY, policy = ReferencePolicy.STATIC)
     private volatile TableStorageStates myTableStorageStates;
 
+    @Reference(service = TableRepairPolicy.class, cardinality = ReferenceCardinality.AT_LEAST_ONE, policy = ReferencePolicy.STATIC)
+    private volatile List<TableRepairPolicy> myRepairPolicies;
+
+
     private volatile RepairSchedulerImpl myDelegateRepairSchedulerImpl;
 
     @Activate
@@ -80,6 +79,7 @@ public class RepairSchedulerService implements RepairScheduler
                 .withRepairStateFactory(myRepairStateFactory)
                 .withRepairLockType(configuration.repairLockType())
                 .withTableStorageStates(myTableStorageStates)
+                .withRepairPolicies(myRepairPolicies)
                 .build();
     }