+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/TestTimeBasedRunPolicy.java
@@ -116,6 +116,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
 
         assertThat(delay).isLessThanOrEqualTo(expectedHighest);
         assertThat(delay).isGreaterThan(-1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
     }
 
     @Test
@@ -129,6 +130,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry(keyspace, table, now.minusHours(2), now.minusHours(1));
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(-1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isTrue();
     }
 
     @Test
@@ -142,6 +144,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry(keyspace, table, now.plusHours(1), now.plusHours(2));
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(-1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isTrue();
     }
 
     @Test
@@ -162,6 +165,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
 
         assertThat(delay).isLessThanOrEqualTo(expectedHighest);
         assertThat(delay).isGreaterThan(-1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
     }
 
     @Test
@@ -183,6 +187,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
 
         assertThat(delay).isLessThanOrEqualTo(expectedHighest);
         assertThat(delay).isGreaterThanOrEqualTo(expectedLowest);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
     }
 
     @Test
@@ -196,6 +201,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry(keyspace, table, now.plusHours(2), now.minusHours(1));
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(-1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isTrue();
     }
 
     @Test
@@ -209,9 +215,11 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry("*", table, 0, 0, 0, 0);
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace1, table))).isFalse();
 
         when(myRepairJobMock.getTableReference()).thenReturn(new TableReference(keyspace2, table));
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace2, table))).isFalse();
     }
 
     @Test
@@ -224,6 +232,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry(keyspace, table, 0, 0, 0, 0);
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
     }
 
     @Test
@@ -236,6 +245,7 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry("*", "*", 0, 0, 0, 0);
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
     }
 
     @Test
@@ -248,15 +258,18 @@ public class TestTimeBasedRunPolicy extends AbstractCassandraTest
         insertEntry("*", "*", 0, 0, 0, 0);
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
 
         mySession.execute(String.format("DELETE FROM %s.%s WHERE keyspace_name = '*' AND table_name = '*'", myKeyspaceName, TABLE_REJECT_CONFIGURATION));
 
         assertThat(myRunPolicy.validate(myRepairJobMock)).isEqualTo(DEFAULT_REJECT_TIME);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isFalse();
 
         // Cache expires after 1 sec
         await().pollInterval(500, TimeUnit.MILLISECONDS)
                 .atMost(2, TimeUnit.SECONDS)
                 .until(() -> myRunPolicy.validate(myRepairJobMock) == -1L);
+        assertThat(myRunPolicy.shouldRun(new TableReference(keyspace, table))).isTrue();
     }
 
     @Test